ウェアラブルのコードからheaderState排除
